fix: make WithHistoryTrimmedTo an internal extension method
Fixes #95
