+++ src/A2A/Server/TaskManager.cs
@@ -1,3 +1,4 @@
+using A2A.Extensions;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
